+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased
+## [Unreleased]
+### Fixed
+ - Resolve error when running `py.test --fixtures`
 
 
 ## [1.1.0] - 2020-08-26